Dependency Inversion (DIP)
Modules/classes should depend on abstractions.

Many classes were already made dependent
on abstract classes and not concrete ones
(constructors take reference or pointer to an
interface/polymorphic base) in the previous commits.

So in this commit created a separate directory
for common interfaces. Moved the interfaces to that
directory or to directories of their clients.

Because under DIP clients/users of an interface
usually keep it and other modules implement it.
If there are many clients/users, the interface
is usually kept in a common directory.

Thus, the project is now in accordance with DIP.